Fix NaN Error on Zero Vector and Refactor
If the displacement vector is a zero vector, its magnitude will be zero and
cause the result of the division be infinity. Passing infinity to SEAL is an
undefined behaviour and thus will result the position of the listener being
NaN.

The thread for moving the listener is better named by `walker' rather than
`mover'.