Fix thread sync and refactor
When cleaning up the queue, the updater thread must be stoppped (joined) before
hand in order to prevent it from pushing more stuff into the queue. Thus a
call to stop_then_clean_queue() or restart_queuing() stops the updater.
The call to seal_rewind_stream() in restart_queuing() is fine because
stop_then_clean_queue() stops the updater before the call.

Now restart_queuing() stops the updater, seal_play_src() only needs to
call wait4updater() in the other case where `src' is not playing. In any
case, the updater will be stopped and the call to _seal_create_thread() no
longer needs a check.