Improve error handling and new/delete style
Returning an error code is a more thread-safe way of handling error in C
than using thread-local storage, which has certain overhead.

It is also good practice to make users of the library capable of
allocating objects themselves and use initialization and finalization
functions to manage the life cycle of objects. It is particularly useful
when they want to use an automatic variable, which is usually on the stack
and cannot be used the old way, where all the data types are opaque
(abstract).

Whether there will be support for MIDI is not finalized yet, so there is
no point of keeping place holder files in the working tree.