Remove OpenAL global lock
The OpenAL global lock is added for the thread safety of OpenAL's global
error state, but it turns out that OpenAL (at least the OpenAL Soft
implementation of it) uses thread-local context, and therefore the error
of each context is also thread-local. There is no need to lock OpenAL
calls if the error state is thread-local.