Add custom directory timestamp target
Since the default timestamp of a directory changes whenever the content of
the directory changes, it is not a good idea to make any target depend on
the directory itself as the latter will always be newer than the former. A
user-touched file under the directory can be used for this specific
purpose.