Fix module documentations
YARD parser doesn't execute code and trace function calls, so helper
functions that defines a moduler under another module is not properly
parented. Making them all under one function solves this as every
variable is local to the function).

Also, `Document-class` should be replaced to `Document-module` when
documenting a module.