Remove dependency on `Seal.startup` for reset
Since listener is a singleton, tests that alters its state will affect
other tests. Currently this is avoided by calling `Seal.startup` for
every test in the `by default` examples, but the readability is low. We
can prevent the test pollution by reseting the state every time we
change a state.