Use more meaningful stack trace when raising error
By default the top of the call stack will be the check_error() call frame,
which is a generic helper for checking Seal API errors, and is not useful.
`caller.shift` toss the first frame away and make the top of stack
pointing to the line where the Seal API call happens.